+++ GHEDT/__init__.py
@@ -1,2 +1,0 @@
-from . import PLAT
-from . import ground_heat_exchangers